docs/pyproject.toml
[project]
name = "documentacion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "furo>=2024.8.6",
    "myst-parser>=4.0.1",
    "pygments-csv-lexer>=0.1.3",
    "sphinx>=7.0.0",
    "sphinx-copybutton>=0.5.2",
    "sphinxcontrib-bibtex>=2.6.3",
]
